Fixes the RSS + Sitemap solution - of course I can't use fs... /facepalm.
