+++ app/routes/[sitemap.xml].tsx
@@ -1,12 +1,9 @@
 
 import { LoaderArgs } from "@remix-run/node";
-import fsAsync from "fs/promises";
-import { getSitemaps } from "utils/generateSitemaps";
+import { getSitemaps } from "utils/getSitemap";
 
 export async function loader({request} : LoaderArgs) {
     const sitemap = getSitemaps();
-    
-
     return new Response(sitemap, {
         headers: {
             "Content-Type": "application/xml",
