+++ package.json
@@ -2,7 +2,7 @@
   "private": true,
   "sideEffects": false,
   "scripts": {
-    "build": "npm run generate:mdxjson && npm run generate:sitemaps && npm run generate:rss && remix build --sourcemap && sentry-upload-sourcemaps --org black-sheep-code --project javascript-remix",
+    "build": "npm run generate:mdxjson && remix build --sourcemap && sentry-upload-sourcemaps --org black-sheep-code --project javascript-remix",
     "dev": "remix dev",
     "start": "netlify serve",
     "lint": "eslint .",
@@ -10,9 +10,6 @@
     "test:cypress": "cypress open",
     "test:cypress:ci": "cypress run",
     "generate:mdxjson": "rm -rf app/generated/frontmatter && tsx utils/extractFrontMatter.bin.ts",
-    "generate:sitemaps": "tsx utils/generateSitemaps.ts",
-    "generate:rss": "tsx utils/generateRss.ts",
-
     "test": "jest --watchAll"
   },
   "dependencies": {
